+++ lib/widget/big_button.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 class BigButton extends StatelessWidget {
@@ -16,7 +15,6 @@ class BigButton extends StatelessWidget {
   Widget build(BuildContext context) {
     return ElevatedButton(
       style: ElevatedButton.styleFrom(
-        padding: EdgeInsets.symmetric(vertical: defaultTargetPlatform == TargetPlatform.windows ? 20 : 10),
         shape: RoundedRectangleBorder(
           borderRadius: BorderRadius.circular(10),
         ),